fix: wgpu 29.0.3 API fixes in gfx.rs, post.rs, render.rs; silt-lua 0.3 fixes in error.rs, bundle.rs, lua_define.rs, lua_ent.rs, model.rs, types.rs
Agent-Logs-Url: https://github.com/Auxnon/Petrichor64/sessions/646f6d9c-2fe4-40ab-952d-cdbf4124bace

+++ src/model.rs
@@ -562,7 +562,7 @@ impl ModelManager {
                 let mut v2 = v.clone();
                 // v2.trans(offset.clone());
                 #[cfg(feature = "headed")]
-                v2.texture(uv[(i / 4 as usize)]);
+                v2.texture(uv[i / 4usize]);
                 v2
             })
             .collect::<Vec<Vertex>>();
